Extract RecordWorkflowTaskStartedRequest handler
